refactor: reorganize node sync tests and remove obsolete files
- Delete outdated investigation and test files related to the node sync concurrency issue.
- Update imports in existing test files to reference the new file logger location.
- Introduce a new file logger implementation to enhance logging capabilities during tests.

These changes streamline the test suite and improve the overall organization of the node sync testing infrastructure.
