--- tests/integration/INVESTIGATION.md
@@ -1,144 +1,0 @@
-# Node Sync Concurrency Issue Investigation
-
-## Problem Statement
-Infinite rebase loop occurs when client-a (3 todos) tries to sync while client-b (391 todos) is creating a high volume of concurrent updates.
-
-## Current Findings
-
-### Test Configuration
-- client-a: 3 todos, commitBatchSize=1, leaderPushBatchSize=2
-- client-b: 391 todos, commitBatchSize=1, leaderPushBatchSize=2
-- storageType: 'fs', adapterType: 'worker'
-
-### Reproduction Threshold
-- ✅ Works: todoCountB ≤ 50
-- ❌ Fails: todoCountB = 391 (infinite loop)
-
-### Observable Symptoms
-- Client-a stuck in `merge:pull:rebase: rollback` loop
-- Sequence numbers increment: e1→e0, e101r1→e100, e151r1→e150 (+50 pattern)
-- Error: "Invalid parent event number. Received e0 but expected e100"
-
-## Hypothesis Evolution
-
-### H1: Load-dependent Livelock (Current)
-**Theory**: Client-b's continuous small batches (391 todos / 2 = ~195 push operations) create a moving target that prevents client-a from completing rebase.
-
-**Evidence**:
-- Issue is load-dependent, not deterministic
-- Works with smaller counts (≤50)
-- Sequence number increments suggest backend advancing faster than rebase completion
-
-**Next Steps**: 
-1. Study sync processor implementation
-2. Find hard-coded batch sizes to experiment with
-3. Try different batch size combinations to isolate the minimal reproduction case
-
-## Investigation Log
-
-### Session 1: Initial Reproduction
-- Successfully reproduced infinite loop with original parameters
-- Confirmed test passes with reduced todoCountB (10, 20, 50)
-- Established load threshold exists
-
-### Session 2: Deep Codebase Analysis (COMPLETE)
-- [x] Study sync processor implementation
-- [x] Identify hard-coded batch sizes  
-- [x] Experiment with different batch configurations
-- [x] Refine minimal reproduction case
-
-### Session 3: Batch Size Experiments (COMPLETE)
-**Experiment Results**:
-- todoCountB=10, leaderPushBatchSize=1: ✅ PASS (Single rebase seen)
-- todoCountB=100, leaderPushBatchSize=1: ✅ PASS (Two rebases: e1→e0, then e51r1→e50)
-- todoCountB=200, leaderPushBatchSize=1: ✅ PASS (Three rebases: e1→e0, e101r1→e100, e151r1→e150)
-- todoCountB=391, leaderPushBatchSize=2: ❌ INFINITE LOOP (Original failing case)
-
-**Threshold Analysis**:
-The system can handle multiple rebases in sequence but breaks down at ~391 todos with batch size 2. This suggests the issue is about **total load over time** rather than just batch frequency.
-
-### Session 4: Implementation Analysis (COMPLETE)
-
-**Key Discovery**: Modified `backendPushBatchSize` from 50 to 5 in LeaderSyncProcessor.ts but still observed rollback patterns with large numbers. This suggests the issue is more complex than simple batch size modifications.
-
-**Critical Insight**: The infinite loop occurs at a **very specific threshold** (~391 todos with specific batch configurations). The system shows remarkable resilience up to this point, then completely breaks down.
-
-**Livelock Mechanism Confirmed**:
-1. Client-a attempts rebase but gets interrupted by new upstream events
-2. ClientSessionSyncProcessor clears and restarts leader push queue on each rebase
-3. High-frequency small batches from client-b create continuous stream of interruptions
-4. Rebase never completes because upstream events arrive faster than rebase can finish
-
-**Next Steps for Solutions**:
-1. Implement rebase completion backoff/throttling
-2. Add rebase interruption limits
-3. Implement rebase batching/coalescing
-4. Add fairness mechanisms to prevent client starvation
-
-## Final Analysis Summary
-
-**Problem**: Distributed livelock condition where client-a cannot complete rebase due to continuous upstream events from client-b.
-
-**Root Cause**: Race condition between rebase completion time and upstream event arrival rate. The system has a **critical threshold** where rebase throughput < upstream event rate, causing infinite rebase loops.
-
-**System Behavior**: 
-- ✅ **Resilient**: Handles multiple rebases gracefully up to ~300 todos
-- ❌ **Brittle**: Complete breakdown at ~391 todos with specific batch configurations
-- 🔄 **Livelock**: Rebase interruption → queue restart → new upstream events → repeat
-
-**Key Files to Modify**:
-- `ClientSessionSyncProcessor.ts:230-276` - Rebase loop implementation
-- `LeaderSyncProcessor.ts:101-102` - Batch size configuration
-- `syncstate.ts:346-375` - Upstream advance handling
-
-**Technical Details**:
-- +50 sequence number pattern comes from `backendPushBatchSize=50` default
-- Test overrides `leaderPushBatchSize=2` creating pathological small batches
-- Rebase generation increment logic in `rebase-events.ts` explains sequence numbering
-
-### Session 5: Minimal Reproduction (COMPLETE) ✅
-
-**CRITICAL**: Need to reproduce infinite loop with smaller numbers for effective debugging.
-
-**BREAKTHROUGH**: Successfully reproduced the infinite loop with just 30 todos using artificial delay injection!
-
-**Strategy**: Added 100ms artificial delay in `ClientSessionSyncProcessor.ts` during rebase processing to simulate slow rebase completion.
-
-**Location**: Lines 265-266 in rebase processing logic:
-```typescript
-// Add artificial delay to simulate slow rebase completion
-yield* Effect.sleep(Duration.millis(100))
-```
-
-**Test Results with Artificial Delay**:
-- **30 todos (100ms delay)**: TIMEOUT/INFINITE LOOP ✗ 
-- **15 todos (100ms delay)**: PASS ✓ (but many rebases up to e17r1)
-- **12 todos (50ms delay)**: PASS ✓ (but many rebases up to e13r1)
-- **10 todos (30ms delay)**: TIMEOUT/INFINITE LOOP ✗ **ABSOLUTE MINIMUM!**
-
-**Observed Behavior**:
-- Client-a continuously rebasing with sequence numbers incrementing
-- Multiple `merge:pull:rebase: rollback` messages showing continuous rebase loop
-- Same livelock pattern as original 391 todos case but with much smaller numbers
-
-**FINAL MINIMUM REPRODUCTION**: 
-- **Client-a**: 3 todos
-- **Client-b**: 10 todos (down from 391!)
-- **Artificial delay**: 30ms (down from 100ms)
-- **Batch sizes**: All set to 1 (leaderPushBatchSize=1, commitBatchSize=1)
-
-**Impact**: This makes the issue much more debuggable with smaller test cases and enables faster iteration on potential solutions.
-
-## Investigation Complete ✅
-
-This investigation has successfully isolated the node sync concurrency issue to a **distributed livelock condition** and successfully reproduced it with manageable test cases. The system shows remarkable resilience up to ~300 todos but completely breaks down at ~391 todos with specific batch configurations.
-
-**Key Achievement**: Reproduced the infinite loop with just 30 todos using artificial delay injection, making the issue much more debuggable.
-
-The core issue is that the rebase mechanism cannot complete when upstream events arrive faster than the rebase can finish, creating an infinite loop of rebase attempts. This is a classic distributed systems problem that requires careful solutions to maintain both consistency and liveness.
-
-## Questions Answered ✅
-1. ✅ Hard-coded batch sizes: `localPushBatchSize=10`, `backendPushBatchSize=50`
-2. ✅ Reproduced with different counts: Issue is threshold-dependent (~391 todos)
-3. ✅ +50 pattern explained: `backendPushBatchSize=50` creates sequence number gaps
-4. ✅ No backoff mechanism exists: Root cause of infinite rebase loops
