+++ tests/integration/src/tests/node-sync/fixtures/vitest-sync-setup.ts
@@ -5,11 +5,13 @@ import { afterAll, beforeAll } from 'vitest'
 
 let wranglerProcess: ChildProcess.ChildProcess
 
+// TODO we should rewrite this with Effect
 beforeAll(async () => {
   const syncPort = await getFreePort.pipe(Effect.runPromise)
 
   process.env.LIVESTORE_SYNC_PORT = syncPort.toString()
 
+  // TODO it still seems to be the cast that the wrangler process gets leaked somehow. We need to fix this.
   // console.log(`Starting sync backend via \`wrangler dev\` on localhost:${syncPort}`)
   wranglerProcess = ChildProcess.spawn('bunx', ['wrangler', 'dev', '--port', syncPort.toString()], {
     stdio: ['ignore', 'pipe', 'pipe'], // ignore stdin, pipe stdout and stderr to parent process
