+++ tests/integration/src/tests/node-sync/livestore.worker.ts
@@ -4,7 +4,7 @@ import { IS_CI } from '@livestore/utils'
 import { Effect } from '@livestore/utils/effect'
 import { OtelLiveDummy, PlatformNode } from '@livestore/utils/node'
 import { OtelLiveHttp } from '@livestore/utils-dev/node'
-import { makeFileLogger } from './file-logger.js'
+import { makeFileLogger } from './fixtures/file-logger.js'
 import { schema } from './schema.js'
 
 const argv = getWorkerArgs()
